requirements.txt
fastapi
uvicorn
requests
pydantic
pandas
beautifulsoup4
markdown
Pillow